// src/app/models/colaboradores.models.ts
export interface Colaborador {
  tipo:number;
  idEleccion:number;
  idProceso:number;
  tipoDocumento: string;
  numeroDocumento: string;
  nombre: string;
  apellidoPaterno: string;
  apellidoMaterno: string;
  cargo: string;
  sede: string;
  //fecharegisto: Date;
  emaildifusion: string;
  estado :number;
  //idusuario: number;
}